fix(supply-chain): progressive audit tab UX and actionable findings
Route audit runs to the Audit tab with step progress, move the package workbench there with scrollable paginated findings and modal detail, filter informational unknown-package noise, discover nested workspace manifests, and keep audit phase transitions stable through completion.
