+++ src/codex_plugin_scanner/guard/daemon/static/assets/chunks/about-workspace.js
@@ -1,4 +1,4 @@
-import { j as jsxRuntimeExports, ac as Tag, aX as Surface, B as Badge, r as reactExports, S as SectionLabel, aY as HiMiniArrowTopRightOnSquare, aZ as HiMiniCheckBadge } from "../guard-dashboard.js";
+import { j as jsxRuntimeExports, ac as Tag, bl as Surface, B as Badge, r as reactExports, S as SectionLabel, aM as HiMiniArrowTopRightOnSquare, bm as HiMiniCheckBadge } from "../guard-dashboard.js";
 const ABOUT_PARTNER_SECTION_TITLE = "Standards partner program";
 const ABOUT_PARTNER_SECTION_BODY = "Join teams building on HOL open standards. Partners get early access to protocol drafts, co-marketing, and direct engineering support.";
 const ABOUT_PARTNER_CTA = "Explore partner programs";
