+++ src/codex_plugin_scanner/guard/daemon/static/assets/chunks/policy-workspace-page.js
@@ -1,4 +1,4 @@
-import { j as jsxRuntimeExports, S as SectionLabel, o as HiMiniXMark, B as Badge, ac as Tag, aA as HiMiniCommandLine, w as HiMiniExclamationTriangle, aD as scopeLabel, h as harnessDisplayName, A as ActionButton, aE as guardAwareHref, m as formatRelativeTime, aF as HiMiniDocumentText, d as HiMiniCheckCircle, aG as HiMiniCloudArrowUp, aH as HiMiniCheck, aI as HiMiniCodeBracket, aJ as HiMiniClipboardDocument, aK as HiMiniUsers, aL as HiMiniBeaker, aM as HiMiniFolder, aN as HiMiniInformationCircle, Q as HiMiniLockClosed, l as HiMiniShieldCheck, r as reactExports, aO as createCloudExceptionRequest, b as EmptyState, ad as HiMiniMagnifyingGlass, p as HiMiniChevronUp, q as HiMiniChevronDown, y as HiMiniChevronRight, aP as HiMiniPuzzlePiece, aQ as HiMiniGlobeAlt, aR as HiMiniClock, aS as policyActionLabel, aT as fetchCloudExceptions, aU as fetchCloudExceptionRequests, ax as HiMiniTrash, aV as HiMiniCube, aw as HiMiniArrowPath, t as HiMiniCloud, Y as fetchSettings, _ as updateSettings, aW as HiMiniQueueList, aX as HiMiniNoSymbol, aY as HiMiniArrowRight, aZ as HiMiniPlay, aB as WorkspacePageHeader, aC as __vitePreload } from "../guard-dashboard.js";
+import { j as jsxRuntimeExports, S as SectionLabel, o as HiMiniXMark, B as Badge, ac as Tag, aA as HiMiniCommandLine, w as HiMiniExclamationTriangle, b0 as scopeLabel, h as harnessDisplayName, A as ActionButton, b1 as guardAwareHref, m as formatRelativeTime, b2 as HiMiniDocumentText, d as HiMiniCheckCircle, b3 as HiMiniCloudArrowUp, b4 as HiMiniCheck, b5 as HiMiniCodeBracket, b6 as HiMiniClipboardDocument, b7 as HiMiniUsers, aG as HiMiniBeaker, b8 as HiMiniFolder, b9 as HiMiniInformationCircle, Q as HiMiniLockClosed, l as HiMiniShieldCheck, r as reactExports, ba as createCloudExceptionRequest, b as EmptyState, ad as HiMiniMagnifyingGlass, p as HiMiniChevronUp, q as HiMiniChevronDown, y as HiMiniChevronRight, bb as HiMiniPuzzlePiece, bc as HiMiniGlobeAlt, aE as HiMiniClock, bd as policyActionLabel, be as fetchCloudExceptions, bf as fetchCloudExceptionRequests, ax as HiMiniTrash, bg as HiMiniCube, aw as HiMiniArrowPath, t as HiMiniCloud, Y as fetchSettings, _ as updateSettings, bh as HiMiniQueueList, bi as HiMiniNoSymbol, bj as HiMiniArrowRight, bk as HiMiniPlay, a_ as WorkspacePageHeader, a$ as __vitePreload } from "../guard-dashboard.js";
 const CLOUD_EXCEPTION_EXPIRING_SOON_DAYS = 7;
 function parseCloudExceptionTimestamp(value) {
   if (!value || !value.trim()) {
