+++ src/codex_plugin_scanner/guard/daemon/static/assets/guard-dashboard.js
@@ -25951,7 +25951,7 @@ const SettingsWorkspace = reactExports.lazy(() => __vitePreload(() => import("./
 const AppDetailWorkspace = reactExports.lazy(() => __vitePreload(() => import("./chunks/app-detail-workspace.js"), true ? [] : void 0).then((m) => ({ default: m.AppDetailWorkspace })));
 const HelpModal = reactExports.lazy(() => __vitePreload(() => import("./chunks/help-modal.js"), true ? [] : void 0).then((m) => ({ default: m.HelpModal })));
 const SupplyChainHubWorkspace = reactExports.lazy(
-  () => __vitePreload(() => import("./chunks/supply-chain-hub-workspace.js").then((n) => n.s), true ? [] : void 0).then((m) => ({ default: m.SupplyChainHubWorkspace }))
+  () => __vitePreload(() => import("./chunks/supply-chain-hub-workspace.js").then((n) => n.a), true ? [] : void 0).then((m) => ({ default: m.SupplyChainHubWorkspace }))
 );
 const PolicyWorkspacePage = reactExports.lazy(
   () => __vitePreload(() => import("./chunks/policy-workspace-page.js"), true ? [] : void 0).then((m) => ({ default: m.PolicyWorkspacePage }))
@@ -26664,7 +26664,7 @@ export {
   fetchRuntimeSnapshot as Z,
   updateSettings as _,
   EvidenceActivityHeatmapMini as a,
-  HiMiniArrowTopRightOnSquare as a$,
+  __vitePreload as a$,
   clearReviewQueue as a0,
   revokeApprovalGateCooldown as a1,
   disableApprovalGateTotp as a2,
@@ -26676,32 +26676,32 @@ export {
   exportSettings as a8,
   importSettings as a9,
   HiMiniCommandLine as aA,
-  WorkspacePageHeader as aB,
-  __vitePreload as aC,
-  scopeLabel as aD,
-  guardAwareHref as aE,
-  HiMiniDocumentText as aF,
-  HiMiniCloudArrowUp as aG,
-  HiMiniCheck as aH,
-  HiMiniCodeBracket as aI,
-  HiMiniClipboardDocument as aJ,
-  HiMiniUsers as aK,
-  HiMiniBeaker as aL,
-  HiMiniFolder as aM,
-  HiMiniInformationCircle as aN,
-  createCloudExceptionRequest as aO,
-  HiMiniPuzzlePiece as aP,
-  HiMiniGlobeAlt as aQ,
-  HiMiniClock as aR,
-  policyActionLabel as aS,
-  fetchCloudExceptions as aT,
-  fetchCloudExceptionRequests as aU,
-  HiMiniCube as aV,
-  HiMiniQueueList as aW,
-  HiMiniNoSymbol as aX,
-  HiMiniArrowRight as aY,
-  HiMiniPlay as aZ,
-  Surface as a_,
+  isSupplyChainAuditIncomplete as aB,
+  readString$1 as aC,
+  isRecord$2 as aD,
+  HiMiniClock as aE,
+  IconActionButton as aF,
+  HiMiniBeaker as aG,
+  ActivationSummary as aH,
+  ActionResultPanel as aI,
+  HiMiniBugAnt as aJ,
+  GuardModalLayer as aK,
+  ConnectFlowCard as aL,
+  HiMiniArrowTopRightOnSquare as aM,
+  HiMiniCloudArrowDown as aN,
+  fetchPackageFirewallStatus as aO,
+  runPackageAudit as aP,
+  resolveSupplyChainAuditFailure as aQ,
+  runPackageSync as aR,
+  startPackageFirewallConnect as aS,
+  openPackageFirewallAuthorizeWindow as aT,
+  PACKAGE_FIREWALL_CONNECT_POPUP_BLOCKED_MESSAGE as aU,
+  runPackageFirewallAction as aV,
+  parseInterceptProofSnapshot as aW,
+  openPackageFirewallShell as aX,
+  EntitlementNotice as aY,
+  fetchReceipts as aZ,
+  WorkspacePageHeader as a_,
   resetSettings as aa,
   setupDesktopNotifications as ab,
   Tag as ac,
@@ -26729,37 +26729,38 @@ export {
   clearLabelForScope as ay,
   formatHarnessCommand as az,
   EmptyState as b,
-  HiMiniCheckBadge as b0,
-  isSupplyChainAuditIncomplete as b1,
-  readString$1 as b2,
-  isRecord$2 as b3,
-  IconActionButton as b4,
-  ActivationSummary as b5,
-  ActionResultPanel as b6,
-  HiMiniBugAnt as b7,
-  GuardModalLayer as b8,
-  ConnectFlowCard as b9,
-  HiMiniCloudArrowDown as ba,
-  fetchPackageFirewallStatus as bb,
-  runPackageAudit as bc,
-  resolveSupplyChainAuditFailure as bd,
-  runPackageSync as be,
-  startPackageFirewallConnect as bf,
-  openPackageFirewallAuthorizeWindow as bg,
-  PACKAGE_FIREWALL_CONNECT_POPUP_BLOCKED_MESSAGE as bh,
-  runPackageFirewallAction as bi,
-  parseInterceptProofSnapshot as bj,
-  openPackageFirewallShell as bk,
-  EntitlementNotice as bl,
-  fetchSupplyChainBundle as bm,
-  HiMiniDocumentMagnifyingGlass as bn,
-  HiMiniShieldExclamation as bo,
-  HiMiniComputerDesktop as bp,
-  HiMiniArrowDown as bq,
-  HiMiniArrowUp as br,
-  fetchReceipts as bs,
-  runAuditRemediation as bt,
-  HiMiniSignal as bu,
+  scopeLabel as b0,
+  guardAwareHref as b1,
+  HiMiniDocumentText as b2,
+  HiMiniCloudArrowUp as b3,
+  HiMiniCheck as b4,
+  HiMiniCodeBracket as b5,
+  HiMiniClipboardDocument as b6,
+  HiMiniUsers as b7,
+  HiMiniFolder as b8,
+  HiMiniInformationCircle as b9,
+  createCloudExceptionRequest as ba,
+  HiMiniPuzzlePiece as bb,
+  HiMiniGlobeAlt as bc,
+  policyActionLabel as bd,
+  fetchCloudExceptions as be,
+  fetchCloudExceptionRequests as bf,
+  HiMiniCube as bg,
+  HiMiniQueueList as bh,
+  HiMiniNoSymbol as bi,
+  HiMiniArrowRight as bj,
+  HiMiniPlay as bk,
+  Surface as bl,
+  HiMiniCheckBadge as bm,
+  fetchSupplyChainBundle as bn,
+  HiMiniDocumentMagnifyingGlass as bo,
+  HiMiniShieldExclamation as bp,
+  HiMiniComputerDesktop as bq,
+  HiMiniChevronLeft as br,
+  HiMiniArrowDown as bs,
+  HiMiniArrowUp as bt,
+  runAuditRemediation as bu,
+  HiMiniSignal as bv,
   EvidenceInsightsShareModal as c,
   HiMiniCheckCircle as d,
   GuardHero as e,
